requirements.txt
streamlit
python-dotenv
langchain  
langchain_community
langchain_core
openai
pypdf
pdfplumber